make base_payment_rule_id also int64

+++ publisher.go
@@ -20,7 +20,7 @@ type Publisher struct {
 	Name                  string `json:"name"`
 	IsOO                  bool   `json:"is_oo,omitempty"`
 	ResellingExposure     string `json:"reselling_exposure,omitempty"`
-	BasePaymentRuleID     int    `json:"base_payment_rule_id,omitempty"`
+	BasePaymentRuleID     int64  `json:"base_payment_rule_id,omitempty"`
 	InventoryRelationship string `json:"inventory_relationship,omitempty"`
 	InventorySource       string `json:"inventory_source,omitempty"`
 }
